% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cded.R
\name{cded_fn}
\alias{cded_fn}
\title{Download Canadian Digital Elevation Data (CDED) intersecting the user provided area of interest.}
\usage{
cded_fn(
  aoi,
  cdedurl = "https://pub.data.gov.bc.ca/datasets/175624/",
  resolution = NULL
)
}
\description{
This function accepts a user provided area of interest as a sf object and retrieves CDED
data \code{https://catalogue.data.gov.bc.ca/dataset/7b4fef7e-7cae-4379-97b8-62b03e9ac83d} for each of the NTS 50K Grid  Digital Baseline Mapping at 1:50,000 (NTS) polygons
that intersect. The CDED data is obtained from \code{"https://pub.data.gov.bc.ca/datasets/175624/"}.
The output CDED will be projected to the coordinate reference system of the area of interest provided,
and returned as a raster object.
}
\details{
@param aoi Area of interest to intersect CDED as sf object.
@param cdedurl (Optional) Web address to CDED data, defaults to \code{"https://pub.data.gov.bc.ca/datasets/175624/"}.
@param resolution (Optional) If provided, CDED will be resampled to the specified resolution using a bilinear method.
@return A raster object containing CDED data intersected and projected to area of interest.
@export
}
