% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDED_fn.R
\name{cded_get}
\alias{cded_get}
\title{Download a CDED DEM}
\usage{
cded_get(aoi)
}
\arguments{
\item{aoi}{An area of interest as a simple features object.}
}
\value{
A CDED digital elevation model as a raster object, cropped to area of interest.
}
\description{
This function takes an area of interest \code{aoi} and obtains all intersecting CDED digital elevation tiles from \link{https://pub.data.gov.bc.ca/datasets/175624}.
Output DEM is merged and cropped to the \code{aoi}.
}
